boot: Move platform-specific code of libavb to libavb_user
Tracked-On: OAM-88563
