+++ avb/libavb_user/uefi_avb_ops.h
