+++ avb/libavb_user/uefi_avb_util.c
